% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shakers.R
\name{inspect_shaker}
\alias{inspect_shaker}
\title{Access the original source vector for a given shaker function}
\usage{
inspect_shaker(f)
}
\description{
Access the original source vector for a given shaker function
}
\examples{
inspect_shaker(shaker$punctuation)
}
