% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{p_indices}
\alias{p_indices}
\title{Sample a proportion of indices of a vector}
\usage{
p_indices(x, p)
}
\arguments{
\item{x}{A vector}

\item{p}{A numeric probability between 0 and 1}
}
\value{
An integer vector of indices.
}
\description{
Sample a proportion of indices of a vector
}
