% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salt_replace.R
\name{salt_character_ocr}
\alias{salt_character_ocr}
\title{Salt a character vector with common OCR problems}
\usage{
salt_character_ocr(x, p = 0.1, ocr_errors = ocr_errors)
}
\arguments{
\item{x}{A character vector}

\item{p}{A number}
}
\description{
Salt a character vector with common OCR problems
}
