% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salt_delete.R
\name{salt_na}
\alias{salt_na}
\alias{salt_empty}
\title{Remove entire values from a vector}
\usage{
salt_na(x, p = 0.2)

salt_empty(x, p = 0.2)
}
\arguments{
\item{x}{A vector}

\item{p}{A number between 0 and 1. Proportion of values to edit.}
}
\value{
A vector the same length as \code{x}
}
\description{
Remove entire values from a vector
}
