% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salty-package.R
\docType{package}
\name{salty-package}
\alias{salty-package}
\title{salty: Turn Clean Data into Messy Data}
\description{
Take real or simulated data and salt it with errors commonly found in the wild, such as pseudo-OCR errors, Unicode problems, numeric fields with nonsensical punctuation, bad dates, etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdlincoln/salty}
  \item Report bugs at \url{https://github.com/mdlincoln/salty/issues}
}

}
\author{
\strong{Maintainer}: Matthew Lincoln \email{matthew.d.lincoln@gmail.com} (\href{https://orcid.org/0000-0002-4387-3384}{ORCID})

}
\keyword{internal}
