% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{season}
\alias{season}
\title{Season an entire data frame by salting each of its columns}
\usage{
season(data, p = 0.2, ...)
}
\arguments{
\item{data}{A data.frame}

\item{p}{A number between 0 and 1. Percent of values in \code{x} that should be salted.}

\item{...}{Column name and \code{salt_} function pairs}
}
\value{
A data.frame with the same number of rows and columns as \code{data}
}
\description{
Season an entire data frame by salting each of its columns
}
