% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shakers.R
\docType{data}
\name{shaker}
\alias{shaker}
\alias{replacement_shaker}
\alias{available_shakers}
\title{Get a set of values to use in \code{salt_} functions}
\format{An object of class \code{list} of length 6.}
\usage{
shaker

replacement_shaker

available_shakers()
}
\value{
A sampling function that will be called by \link{salt_insert},
\link{salt_substitute}, or \link{salt_replace}.
}
\description{
\link{shaker} contains various character sets to be added to your data using
\link{salt_insert} and \link{salt_substitute}. \link{replacement_shaker} is for
\link{salt_replace}, and contains pairlists that replace matched patterns in your
data.
}
\examples{
salt_insert(letters, shaker$punctuation)
available_shakers()
}
\keyword{datasets}
